# Conflicts:
#	innerPeace/src/main/java/com/social/innerPeace/repository/PostRepository.java

+++ innerPeace/src/main/java/com/social/innerPeace/board/post/service/BoardPostServiceImpl.java
@@ -22,10 +22,7 @@ import org.springframework.stereotype.Service;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Base64;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -46,6 +43,8 @@ public class BoardPostServiceImpl implements BoardPostService {
     private LikeRepository likeRepository;
     @Autowired
     private FileStore fileStore;
+
+    private final Map<String, String> imageCache = new HashMap<>();
 
     @Override
     public String write(PostDTO dto) {
@@ -88,11 +87,69 @@ public class BoardPostServiceImpl implements BoardPostService {
         return listEntityToDto(posts);
     }
 
+    @Override
+    public List<PostDTO> scrollList(Long postNo) {
+        Pageable pageable = PageRequest.of(0, 36, Sort.by("postNo").descending());
+        Page<Post> posts = postRepository.findByPostNoLessThanOrderByPostNoDesc(postNo, pageable);
+        List<Post> postList = posts.getContent(); // 실제 Post 목록을 얻음
+        return listEntityToDto(postList);
+    }
+
+    @Override
+    public List<PostDTO> search(String searchkey) {
+        String modifiedSearchKey = searchkey.trim(); // 앞뒤 공백 제거
+
+        // searchKey가 "#"으로 시작하는 경우
+        if (modifiedSearchKey.startsWith("#")) {
+            modifiedSearchKey = modifiedSearchKey.substring(1); // "#" 제거
+            Pageable pageable = PageRequest.of(0, 36, Sort.by("postNo").descending());
+            Page<Post> posts = postRepository.findByTagContaining(modifiedSearchKey, pageable);
+            List<Post> postList = posts.getContent(); // 실제 Post 목록을 얻음
+            return listEntityToDto(postList);
+        }
+        // searchKey가 "@"으로 시작하는 경우
+        else if (modifiedSearchKey.startsWith("@")) {
+            modifiedSearchKey = modifiedSearchKey.substring(1); // "@" 제거
+            Pageable pageable = PageRequest.of(0, 36, Sort.by("postNo").descending());
+            Page<Post> posts = postRepository.findByHealerNickNameContaining(modifiedSearchKey, pageable);
+            List<Post> postList = posts.getContent(); // 실제 Post 목록을 얻음
+            return listEntityToDto(postList);
+        }
+        Pageable pageable = PageRequest.of(0, 36, Sort.by("postNo").descending());
+        Page<Post> posts = postRepository.findByContentContaining(modifiedSearchKey, pageable);
+        List<Post> postList = posts.getContent(); // 실제 Post 목록을 얻음
+        return listEntityToDto(postList);
+    }
+
+    @Override
+    public List<PostDTO> searchScrollList(Long postNo, String search) {
+        String modifiedSearchKey = search.trim(); // 앞뒤 공백 제거
+
+        // searchKey가 "#"으로 시작하는 경우
+        if (modifiedSearchKey.startsWith("#")) {
+            modifiedSearchKey = modifiedSearchKey.substring(1); // "#" 제거
+            Pageable pageable = PageRequest.of(0, 36, Sort.by("postNo").descending());
+            Page<Post> posts = postRepository.findByTagContainingAndPostNoLessThan(modifiedSearchKey, postNo, pageable);
+            List<Post> postList = posts.getContent(); // 실제 Post 목록을 얻음
+            return listEntityToDto(postList);
+        }
+        // searchKey가 "@"으로 시작하는 경우
+        else if (modifiedSearchKey.startsWith("@")) {
+            modifiedSearchKey = modifiedSearchKey.substring(1); // "@" 제거
+            Pageable pageable = PageRequest.of(0, 36, Sort.by("postNo").descending());
+            Page<Post> posts = postRepository.findByHealerNickNameContainingAndPostNoLessThan(modifiedSearchKey, postNo, pageable);
+            List<Post> postList = posts.getContent(); // 실제 Post 목록을 얻음
+            return listEntityToDto(postList);
+        }
+        Pageable pageable = PageRequest.of(0, 36, Sort.by("postNo").descending());
+        Page<Post> posts = postRepository.findByContentContainingAndPostNoLessThan(modifiedSearchKey, postNo, pageable);
+        List<Post> postList = posts.getContent(); // 실제 Post 목록을 얻음
+        return listEntityToDto(postList);
+    }
 
     @Override
     public PostDTO detail(Long postNo, String healerNickName) {
         Post post = postRepository.findByPostNoWithHealer(postNo);
-        String base64String = null;
         if (post != null) {
             List<Post_Like> likeList = likeRepository.findByPostPostNo(postNo);
             List<Follow> followerList = healerRepository.findFollowersByHealerEmail(post.getHealer().getHealerEmail());
@@ -118,14 +175,28 @@ public class BoardPostServiceImpl implements BoardPostService {
                 }
             }
             try {
-                byte[] fileBytes = readBytesFromFile(postimagePath);
-                base64String = encodeBytesToBase64(fileBytes);
+                String base64String = imageCache.get(postimagePath);
+                if (base64String == null) {
+                    byte[] fileBytes = readBytesFromFile(postimagePath);
+                    base64String = encodeBytesToBase64(fileBytes);
+                    imageCache.put(postimagePath, base64String);
+                }
                 postDTO.setPost_image("data:image/png;base64," + base64String);
-                fileBytes = readBytesFromFile(thumbnailImagePath);
-                base64String = encodeBytesToBase64(fileBytes);
+
+                base64String = imageCache.get(thumbnailImagePath);
+                if (base64String == null) {
+                    byte[] fileBytes = readBytesFromFile(thumbnailImagePath);
+                    base64String = encodeBytesToBase64(fileBytes);
+                    imageCache.put(thumbnailImagePath, base64String);
+                }
                 postDTO.setPost_image_thumbnail("data:image/png;base64," + base64String);
-                fileBytes = readBytesFromFile(profileImagePath);
-                base64String = encodeBytesToBase64(fileBytes);
+
+                base64String = imageCache.get(profileImagePath);
+                if (base64String == null) {
+                    byte[] fileBytes = readBytesFromFile(profileImagePath);
+                    base64String = encodeBytesToBase64(fileBytes);
+                    imageCache.put(profileImagePath, base64String);
+                }
                 postDTO.setHealer_profile_image("data:image/png;base64," + base64String);
 
             } catch (IOException e) {
@@ -142,7 +213,7 @@ public class BoardPostServiceImpl implements BoardPostService {
     public PostDTO modify(PostDTO dto, String loginHealer) {
         Optional<Post> optionalPost = postRepository.findById(dto.getPost_no());
         Optional<Healer> optionalHealer = healerRepository.findByHealerNickName(loginHealer);
-        if(!optionalHealer.get().getHealerEmail().equals(optionalPost.get().getHealer().getHealerEmail())){
+        if (!optionalHealer.get().getHealerEmail().equals(optionalPost.get().getHealer().getHealerEmail())) {
             return null;
         }
         if (optionalPost.isPresent()) {
@@ -169,46 +240,16 @@ public class BoardPostServiceImpl implements BoardPostService {
         return null;
     }
 
-    @Override
-    public List<PostDTO> search(String searchkey) {
-        if (searchkey == null) {
-            return null; // searchKey가 null인 경우, null을 반환
-        }
-
-        String modifiedSearchKey = searchkey.trim(); // 앞뒤 공백 제거
-
-        // searchKey가 "#"으로 시작하는 경우
-        if (modifiedSearchKey.startsWith("#")) {
-            modifiedSearchKey = modifiedSearchKey.substring(1); // "#" 제거
-            Pageable pageable = PageRequest.of(0, 36, Sort.by("postNo").descending());
-            Page<Post> posts = postRepository.findByTagContaining(modifiedSearchKey, pageable);
-            List<Post> postList = posts.getContent(); // 실제 Post 목록을 얻음
-            return listEntityToDto(postList);
-        }
-        // searchKey가 "@"으로 시작하는 경우
-        else if (modifiedSearchKey.startsWith("@")) {
-            modifiedSearchKey = modifiedSearchKey.substring(1); // "@" 제거
-            Pageable pageable = PageRequest.of(0, 36, Sort.by("postNo").descending());
-            Page<Post> posts = postRepository.findByHealerNickNameContaining(modifiedSearchKey, pageable);
-            List<Post> postList = posts.getContent(); // 실제 Post 목록을 얻음
-            return listEntityToDto(postList);
-        }
-        Pageable pageable = PageRequest.of(0, 36, Sort.by("postNo").descending());
-        Page<Post> posts = postRepository.findByContentContaining(searchkey, pageable);
-        List<Post> postList = posts.getContent(); // 실제 Post 목록을 얻음
-        return listEntityToDto(postList);
-    }
 
     @Override
-    public int deletePost(PostDTO postDTO) {
-        int result = 0;
-        try{
+    public int deletePost(PostDTO postDTO,String loginHealer) {
+        Optional<Post> optionalPost = postRepository.findById(postDTO.getPost_no());
+        if (!optionalPost.get().getHealer().getHealerNickName().equals(loginHealer)) {
+            return 0;
+        }else{
             postRepository.deleteById(postDTO.getPost_no());
-            result = 1;
-        }catch (Exception e){
-            log.debug(e.getMessage());
+            return 1;
         }
-        return result;
     }
 
     @Override
@@ -239,12 +280,15 @@ public class BoardPostServiceImpl implements BoardPostService {
         for (Post post : posts) {
             String image = findImagename(post.getPostNo()).getPost_image();
             String imagePath = thumbnail_dir + image;
-            String base64String = null;
-            try {
-                byte[] fileBytes = readBytesFromFile(imagePath);
-                base64String = encodeBytesToBase64(fileBytes);
-            } catch (IOException e) {
-                // 예외 처리
+            String base64String = imageCache.get(imagePath);
+            if (base64String == null) {
+                try {
+                    byte[] fileBytes = readBytesFromFile(imagePath);
+                    base64String = encodeBytesToBase64(fileBytes);
+                    imageCache.put(imagePath, base64String);
+                } catch (IOException e) {
+                    // 예외 처리
+                }
             }
             PostDTO dto = PostDTO
                     .builder()
@@ -274,20 +318,11 @@ public class BoardPostServiceImpl implements BoardPostService {
 
     //태그를 나눠서 리스트에 추가하는 메소드
     public static List<String> splitAndClean(String input) {
-        List<String> words = new ArrayList<>();
-
-        // '#'을 기준으로 문자열을 분리
-        String[] splitWords = input.split("#");
-
-        // 분리된 각 단어에서 공백을 제거하여 리스트에 추가
-        for (String word : splitWords) {
-            String cleanWord = word.trim(); // 공백 제거
-            if (!cleanWord.isEmpty()) { // 빈 문자열이 아니면 리스트에 추가
-                words.add("#" + cleanWord);
-            }
-        }
-
-        return words;
+        return Arrays.stream(input.split("#"))
+                .map(String::trim)
+                .filter(word -> !word.isEmpty())
+                .map(word -> "#" + word)
+                .collect(Collectors.toList());
     }
